feat: Rename folder to etl
