+++ src/etl/pdf_and_markdown_pipeline.py
@@ -15,7 +15,7 @@ from langchain_community.document_loaders import UnstructuredMarkdownLoader
 from langchain_core.documents import Document
 import spacy
 
-from src.data_processing.chunks import HEADERS_TO_SPLIT_ON, ChunksFromMarkdow
+from src.etl.chunks import HEADERS_TO_SPLIT_ON, ChunksFromMarkdow
 
 nlp = spacy.load("pt_core_news_md")
 
@@ -33,11 +33,7 @@ class PdfAndMarkdownPipeline:
     chunk_documents: list = field(init=False, default_factory=list)
     chunk_object: ChunksFromMarkdow = field(init=False, default=None)
 
-    def pdf_to_markdown(
-        self,
-        source_file_path: Union[str, Path],
-        dest_dir_path: Union[str, Path]
-    ):
+    def pdf_to_markdown(self, source_file_path: Union[str, Path], dest_dir_path: Union[str, Path]):
         """
         Função responsável por converter arquivos pdf em markdown.
 
@@ -45,11 +41,11 @@ class PdfAndMarkdownPipeline:
             source_file_path (Union[str, Path]): Caminho do arquivo pdf de origem.
             dest_dir_path (Union[str, Path]): Caminho do diretório em que o arquivo markdown gerado deve ser armazenado.
         """
-        
+
         if isinstance(source_file_path, str):
-            source_file_path = Path(source_file_path) 
+            source_file_path = Path(source_file_path)
         if isinstance(dest_dir_path, str):
-            dest_dir_path = Path(dest_dir_path) 
+            dest_dir_path = Path(dest_dir_path)
 
         conv_result = self.docling_converter(source_file_path)
         if conv_result and conv_result.document:
@@ -59,7 +55,6 @@ class PdfAndMarkdownPipeline:
             _log.error("Docling conversion failed or returned an empty document.")
 
         return self
-
 
     def docling_converter(self, source_file_path: Path):
         """
@@ -81,16 +76,18 @@ class PdfAndMarkdownPipeline:
         # pipeline_options.images_scale = IMAGE_RESOLUTION_SCALE
         # pipeline_options.generate_page_images = True
         # pipeline_options.generate_picture_images = True
-        
-        pipeline_options.ocr_options.lang = ["pt", "en", "es",]
+
+        pipeline_options.ocr_options.lang = [
+            "pt",
+            "en",
+            "es",
+        ]
         pipeline_options.accelerator_options = AcceleratorOptions(
             num_threads=4, device=AcceleratorDevice.AUTO
         )
 
         doc_converter = DocumentConverter(
-            format_options={
-                InputFormat.PDF: PdfFormatOption(pipeline_options=pipeline_options)
-            }
+            format_options={InputFormat.PDF: PdfFormatOption(pipeline_options=pipeline_options)}
         )
         ###########################################################################
 
@@ -99,9 +96,8 @@ class PdfAndMarkdownPipeline:
         end_time = time.time() - start_time
 
         _log.info(f"Document {source_file_path} converted in {end_time:.2f} seconds.")
-        
-        return conv_result
 
+        return conv_result
 
     def extract_text_from_docling_document(self, docling_document) -> str:
         """
@@ -128,11 +124,12 @@ class PdfAndMarkdownPipeline:
         file_path = dest_dir_path.joinpath(f"{file_name}.md")
         with open(file_path, "w") as f:
             f.write(text)
-        
+
         _log.info(f"Document {file_path} saved.")
         return self
 
-    def load_markdown(self,
+    def load_markdown(
+        self,
         source_file_path: Union[str, Path],
     ):
         """
@@ -140,11 +137,11 @@ class PdfAndMarkdownPipeline:
 
         Args:
             source_file_path (Union[str, Path]): Caminho do arquivo markdown de origem.
-         
+
         """
 
         if isinstance(source_file_path, str):
-            source_file_path = Path(source_file_path) 
+            source_file_path = Path(source_file_path)
 
         md_loader = UnstructuredMarkdownLoader(source_file_path)
         self.md_documents = md_loader.load()
@@ -152,10 +149,12 @@ class PdfAndMarkdownPipeline:
         return self
 
     def get_chunks(
-            self,
-            chunk_size: int = 400,
-            chunk_overlap: int = 100,
-            headers_to_split_on: List[Tuple[str, str]] = field(default_factory=lambda: HEADERS_TO_SPLIT_ON),
+        self,
+        chunk_size: int = 400,
+        chunk_overlap: int = 100,
+        headers_to_split_on: List[Tuple[str, str]] = field(
+            default_factory=lambda: HEADERS_TO_SPLIT_ON
+        ),
     ):
         """
         Função responsável por gerar os chunks a partir do conteúdo de texto do markdown.
@@ -174,4 +173,3 @@ class PdfAndMarkdownPipeline:
         )
         self.chunk_documents = self.chunk_object.chunk_documents
         return self
-
